refactor(capabilities): :recycle: migrate pwa components to component-store pattern
Applies the Task R migration recipe to all pwa/ components:

- serwist-provider: folder + index.ts (pure, no store needed)
- install-prompt-banner: folder + use-install-prompt-banner-store.ts
  (absorbs useInstallPrompt + usePwaInstallDecision as private sub-hooks)
- layout-additional-content: folderized; useOfflineContactQueue logic
  folded into use-layout-additional-content-store.ts (its only consumer)

Removes old pwa/components/ + pwa/hooks/ two-folder split. Updates
src/app/layout.tsx to import from new barrel paths.
