+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import Head from "next/head";
 import { Viewport } from "next";
 import { ReactNode } from "react";
 import { Open_Sans, Courier_Prime } from "next/font/google";
-import { SerwistProvider } from "@/components/features/pwa/components/serwist-provider";
+import { SerwistProvider } from "@/components/features/pwa/serwist-provider";
 import { LayoutAdditionalContent } from "@/components/design-system/templates/layout-additional-content";
 
 export const openSans = Open_Sans({
