+++ src/components/design-system/templates/layout-additional-content/index.ts
@@ -1,0 +1,1 @@
+export { LayoutAdditionalContent } from "./layout-additional-content";
