+++ src/components/design-system/templates/layout-additional-content/use-layout-additional-content-store.ts
@@ -4,17 +4,22 @@ import { useEffect } from "react";
 import { contactQueue } from "@/lib/background-sync/contact-queue";
 import { trackWith } from "@/lib/tracking/tracking";
 import { tracking } from "@/types/configuration/tracking";
+import { ComponentStore } from "@/types/component-store";
 
-export function useOfflineContactQueue() {
+type LayoutAdditionalContentState = Record<string, never>;
+type LayoutAdditionalContentEffects = Record<string, never>;
+
+export const useLayoutAdditionalContentStore = (): ComponentStore<
+    LayoutAdditionalContentState,
+    LayoutAdditionalContentEffects
+> => {
     useEffect(() => {
         const replayQueue = async () => {
             while (!contactQueue.isEmpty()) {
                 const entry = contactQueue.dequeue();
-
                 if (!entry) {
                     break;
                 }
-
                 try {
                     await fetch("/api/contact", {
                         method: "POST",
@@ -26,7 +31,6 @@ export function useOfflineContactQueue() {
                             honeypot: entry.honeypot,
                         }),
                     });
-
                     trackWith({
                         action: tracking.action.contact_queue_replayed,
                         category: tracking.category.pwa,
@@ -39,13 +43,16 @@ export function useOfflineContactQueue() {
         };
 
         window.addEventListener("online", replayQueue);
-
         if (navigator.onLine) {
             replayQueue();
         }
-
         return () => {
             window.removeEventListener("online", replayQueue);
         };
     }, []);
-}
+
+    return {
+        state: {},
+        effects: {},
+    };
+};
