+++ src/components/features/pwa/install-prompt-banner/index.ts
@@ -1,0 +1,1 @@
+export { InstallPromptBanner } from "./install-prompt-banner";
