+++ src/components/features/pwa/install-prompt-banner/use-install-prompt-banner-store.ts
@@ -1,7 +1,12 @@
 "use client";
 
 import { useEffect, useState, useSyncExternalStore } from "react";
+import { useGlassmorphism } from "@/components/design-system/utils/hooks/use-glassmorphism";
+import { useConsentStore } from "@/components/design-system/utils/hooks/use-consent-store";
 import { writePwaInstallDecision } from "@/lib/pwa/pwa-install-decision";
+import { trackWith } from "@/lib/tracking/tracking";
+import { tracking } from "@/types/configuration/tracking";
+import { ComponentStore } from "@/types/component-store";
 import { usePwaInstallDecision } from "./use-pwa-install-decision";
 
 interface BeforeInstallPromptEvent extends Event {
@@ -9,10 +14,13 @@ interface BeforeInstallPromptEvent extends Event {
     userChoice: Promise<{ outcome: "accepted" | "dismissed" }>;
 }
 
-export interface UseInstallPromptResult {
-    isInstallable: boolean;
-    isInstalled: boolean;
-    promptInstall: () => Promise<void>;
+interface InstallPromptBannerState {
+    glassmorphismClass: string;
+    visible: boolean;
+}
+
+interface InstallPromptBannerEffects {
+    install: () => Promise<void>;
     dismiss: () => void;
 }
 
@@ -26,59 +34,82 @@ const subscribeStandaloneMode = (callback: () => void) => {
 };
 
 const getStandaloneSnapshot = () =>
-    typeof window !== "undefined" &&
-    window.matchMedia("(display-mode: standalone)").matches;
+    typeof window !== "undefined" && window.matchMedia("(display-mode: standalone)").matches;
 
 const getStandaloneServerSnapshot = () => false;
 
-export function useInstallPrompt(): UseInstallPromptResult {
+export const useInstallPromptBannerStore = (): ComponentStore<
+    InstallPromptBannerState,
+    InstallPromptBannerEffects
+> => {
+    const { glassmorphismClass } = useGlassmorphism();
+    const cookieAccepted = useConsentStore();
+    const decision = usePwaInstallDecision();
     const isStandalone = useSyncExternalStore(
         subscribeStandaloneMode,
         getStandaloneSnapshot,
         getStandaloneServerSnapshot,
     );
-    const decision = usePwaInstallDecision();
     const [deferredPrompt, setDeferredPrompt] = useState<BeforeInstallPromptEvent | null>(null);
 
     const isInstalled = isStandalone || decision === "installed";
     const isInstallable = deferredPrompt !== null && decision === null && !isInstalled;
+    const visible = isInstallable && cookieAccepted;
 
     useEffect(() => {
         const handleBeforeInstallPrompt = (e: Event) => {
             e.preventDefault();
             setDeferredPrompt(e as BeforeInstallPromptEvent);
         };
-
         const handleAppInstalled = () => {
             writePwaInstallDecision("installed");
             setDeferredPrompt(null);
         };
-
         window.addEventListener("beforeinstallprompt", handleBeforeInstallPrompt);
         window.addEventListener("appinstalled", handleAppInstalled);
-
         return () => {
             window.removeEventListener("beforeinstallprompt", handleBeforeInstallPrompt);
             window.removeEventListener("appinstalled", handleAppInstalled);
         };
     }, []);
 
-    const promptInstall = async () => {
+    useEffect(() => {
+        if (visible) {
+            trackWith({
+                action: tracking.action.pwa_install_prompt_shown,
+                category: tracking.category.pwa,
+                label: tracking.label.body,
+            });
+        }
+    }, [visible]);
+
+    const install = async () => {
         if (!deferredPrompt) {
             return;
         }
-
         await deferredPrompt.prompt();
         const { outcome } = await deferredPrompt.userChoice;
-
         writePwaInstallDecision(outcome === "accepted" ? "installed" : "dismissed");
         setDeferredPrompt(null);
+        trackWith({
+            action: tracking.action.pwa_install_accepted,
+            category: tracking.category.pwa,
+            label: tracking.label.body,
+        });
     };
 
     const dismiss = () => {
         writePwaInstallDecision("dismissed");
         setDeferredPrompt(null);
+        trackWith({
+            action: tracking.action.pwa_install_dismissed,
+            category: tracking.category.pwa,
+            label: tracking.label.body,
+        });
     };
 
-    return { isInstallable, isInstalled, promptInstall, dismiss };
-}
+    return {
+        state: { glassmorphismClass, visible },
+        effects: { install, dismiss },
+    };
+};
