+++ src/components/features/pwa/serwist-provider/index.ts
@@ -1,0 +1,1 @@
+export { SerwistProvider } from "./serwist-provider";
